requirements.txt
django>=4.1.3,<4.2.0
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2
django-environ
djangorestframework-simplejwt